Move _get_optimization_status to experiment server
